src/main/webapp/app/entities/drug/update/drug-update.component.ts
import { Component, OnInit } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { ActivatedRoute } from '@angular/router';
import { Observable } from 'rxjs';
import { finalize } from 'rxjs/operators';

import { DrugFormService, DrugFormGroup } from './drug-form.service';
import { IDrug } from '../drug.model';
import { DrugService } from '../service/drug.service';

@Component({
  selector: 'jhi-drug-update',
  templateUrl: './drug-update.component.html',
})
export class DrugUpdateComponent implements OnInit {
  isSaving = false;
  drug: IDrug | null = null;

  editForm: DrugFormGroup = this.drugFormService.createDrugFormGroup();

  constructor(protected drugService: DrugService, protected drugFormService: DrugFormService, protected activatedRoute: ActivatedRoute) {}

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ drug }) => {
      this.drug = drug;
      if (drug) {
        this.updateForm(drug);
      }
    });
  }

  previousState(): void {
    window.history.back();
  }

  save(): void {
    this.isSaving = true;
    const drug = this.drugFormService.getDrug(this.editForm);
    if (drug.id !== null) {
      this.subscribeToSaveResponse(this.drugService.update(drug));
    } else {
      this.subscribeToSaveResponse(this.drugService.create(drug));
    }
  }

  protected subscribeToSaveResponse(result: Observable<HttpResponse<IDrug>>): void {
    result.pipe(finalize(() => this.onSaveFinalize())).subscribe({
      next: () => this.onSaveSuccess(),
      error: () => this.onSaveError(),
    });
  }

  protected onSaveSuccess(): void {
    this.previousState();
  }

  protected onSaveError(): void {
    // Api for inheritance.
  }

  protected onSaveFinalize(): void {
    this.isSaving = false;
  }

  protected updateForm(drug: IDrug): void {
    this.drug = drug;
    this.drugFormService.resetForm(this.editForm, drug);
  }
}
